requirements.txt
absl-py>=0.12.0
accelerate>=0.12.0
datasets>=2.0.0  
langdetect==1.0.9
lm_eval==0.4.7
numpy>=1.20.0
sentencepiece
tokenizer
torch>=2.0.0
transformers==4.45.0 
tqdm==4.67.1
tqdm-multiprocess==0.0.11
importlib-metadata>=4.0.0
wandb>=0.12.0  # optional for logging
torchvision